No longer constrain the rendered glyph bitmap
Previously the rendered glyph bitmap was incorrectly constrained by the
requested size (which the API documentation was suggested against), this
was based on the incorrect assumption that the rendered glyph bitmap
will have the same size as the requested size.

This has been fixed.

Although there is a possibility that the unconstrained bitmap may cause
the character preview to be drawn on top of other components, I was yet
able to demonstrace such a case, so the rendering of the bitmap is
currently unconstrained as well. We can fix it later if needed.